backend/backend/requirements.txt
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
pydantic==2.5.2
python-docx
paddlepaddle
paddleocr
phidata
pandas
numpy
opencv-python
sentence_transformers
transformers
pyMuPDF
pillow
ollama
chromadb
python-magic
tiktoken
nltk
xmltodict
markdown
albumentations
